+++ tests/acceptance/fixtures/list/beta.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+function test_beta_one() {
+  assert_true true
+}